refactor(common/ui): extract DealDetailsController to dedupe deal-details handling
Home and Store ViewModels held byte-identical loadDealDetails/dismissDealDetails
implementations (~50 lines each), tripping SonarCloud's 3% duplication-on-new-code
gate at 18.6%. Move the shared state and lifecycle into DealDetailsController in
:common:ui and have each ViewModel delegate.
