+++ feature/home/src/main/java/pm/bam/gamedeals/feature/home/ui/HomeViewModel.kt
@@ -24,7 +24,7 @@ import kotlinx.coroutines.launch
 import pm.bam.gamedeals.common.logFlow
 import pm.bam.gamedeals.common.onError
 import pm.bam.gamedeals.common.ui.deal.DealBottomSheetData
-import pm.bam.gamedeals.common.withMinimumDuration
+import pm.bam.gamedeals.common.ui.deal.DealDetailsController
 import pm.bam.gamedeals.domain.models.Deal
 import pm.bam.gamedeals.domain.models.Giveaway
 import pm.bam.gamedeals.domain.models.Release
@@ -60,12 +60,8 @@ internal class HomeViewModel @Inject constructor(
         initialValue = HomeScreenData()
     )
 
-    private val _dealDetails = MutableStateFlow<DealBottomSheetData?>(null)
-    val dealDetails: StateFlow<DealBottomSheetData?> = _dealDetails.stateIn(
-        scope = viewModelScope,
-        started = SharingStarted.WhileSubscribed(5000),
-        initialValue = null
-    )
+    private val dealDetailsController = DealDetailsController(dealsRepository, storesRepository, logger)
+    val dealDetails: StateFlow<DealBottomSheetData?> = dealDetailsController.dealDetails
 
     private val _events = MutableSharedFlow<HomeUiEvent>(
         replay = 0,
@@ -104,61 +100,12 @@ internal class HomeViewModel @Inject constructor(
                 .collect { _events.emit(HomeUiEvent.NavigateToGame(it)) }
         }
 
-    fun loadDealDetails(
-        dealId: String,
-        dealStoreId: Int,
-        dealTitle: String,
-        dealPriceDenominated: String
-    ) {
-        viewModelScope.launch {
-            try {
-                _dealDetails.emit(
-                    DealBottomSheetData.DealDetailsLoading(
-                        store = storesRepository.getStore(dealStoreId),
-                        gameName = dealTitle,
-                        dealId = dealId,
-                        gameSalesPriceDenominated = dealPriceDenominated
-                    )
-                )
-
-                val data = withMinimumDuration(750L) {
-                    val dealDetails = dealsRepository.getDeal(dealId)
-                    val store = storesRepository.getStore(dealStoreId)
-
-                    DealBottomSheetData.DealDetailsData(
-                        store = store,
-                        gameName = dealTitle,
-                        dealId = dealId,
-                        gameSalesPriceDenominated = dealPriceDenominated,
-                        gameInfo = dealDetails.gameInfo,
-                        cheapestPrice = dealDetails.cheapestPrice,
-                        cheaperStores = dealDetails.cheaperStores.map { storesRepository.getStore(it.storeID) to it }
-                    )
-                }
-                _dealDetails.emit(data)
-            } catch (t: Throwable) {
-                fatal(logger, t)
-                try {
-                    _dealDetails.emit(
-                        DealBottomSheetData.DealDetailsError(
-                            store = storesRepository.getStore(dealStoreId),
-                            gameName = dealTitle,
-                            dealId = dealId,
-                            gameSalesPriceDenominated = dealPriceDenominated
-                        )
-                    )
-                } catch (inner: Throwable) {
-                    fatal(logger, inner)
-                    dismissDealDetails()
-                }
-            }
-        }
+    fun loadDealDetails(dealId: String, dealStoreId: Int, dealTitle: String, dealPriceDenominated: String) {
+        dealDetailsController.load(viewModelScope, dealId, dealStoreId, dealTitle, dealPriceDenominated)
     }
 
     fun dismissDealDetails() {
-        viewModelScope.launch {
-            _dealDetails.emit(null)
-        }
+        dealDetailsController.dismiss(viewModelScope)
     }
 
     private fun loadTopStoreDataFlow() =
